:bug: fix: ゲストアカウント以外が削除されないようにevents.signOutを修正
close #202

+++ src/pages/api/auth/[...nextauth].ts
@@ -59,9 +59,12 @@ export const authOptions: NextAuthOptions = {
       ]);
     },
     async signOut({ token }) {
-      await prisma.user.delete({
-        where: { id: token.sub },
-      });
+      // ゲストアカウントはsignOut時に削除する
+      if (token.email?.endsWith("@guestUser.com")) {
+        await prisma.user.delete({
+          where: { id: token.sub },
+        });
+      }
     },
   },
 };
